Native ABI surface: std.ffi-typed exact-width scalars + Option RawPtr nullability (N-20 redo / N-24 / N-23)
Make the raw host.native surface obey §26.1.1/§26.1.2 conservative typing:

- §26.1.2:26114 (exact-width MUST): the exact-width / pointer-width / float
  classes (ctI8..ctU64, ctIsize, ctUsize, ctF32, and ctInt64 = int64_t) now
  surface as the std.ffi NOMINAL types (I8..U64/Isize/Usize/F32/I64), not bare
  Integer. The marshalling wrappers construct/destructure the MkXxx wrapper
  ctors (kctor/kctor_arg), and codegen collects their KT_ tag ids (shared with
  any in-program match), so construction and destructuring agree. ctInt (C
  `int`, not exact-width) stays the ergonomic Integer surface.
- §26.1.1:27307-27309 (RawPtr nullability MUST): a RawPtr result/param surfaces
  as `Option RawPtr`; marshalling maps NULL↔None and non-null↔Some(MkRawPtr a).
- §27.1.1:27316: bare OpaqueHandle is explicitly permitted for a RAW host
  binding whose release is supplied out of band — the demo's host.native.*
  raw surface + sqliteClose qualifies; documented in NativeFfi (so N-24's
  owning-handle concern is compliant, and N-23 M.Raw is N/A since we expose
  only the raw surface, not a refined one).

std.ffi gains ergonomic constructors (i8..usize, f32, f64, rawPtr) and accessors
(i8Value..f64Value, rawPtrAddr). The demo's sqliteQueryInt returns I64 (unwrapped
via i64Value); tcpListen's port is C `int` (ctInt). native_shim.c updated to match.

Tests: cabal test 242/242 PASS; tests/build PASS; test/native PASS incl. two NEW
end-to-end checks — ctU32→U32 (`htonl`) and ctRawPtr→Option RawPtr (`getenv`→Some,
matched in Kappa) — plus the existing ABI-verify, string-dispatch, and lock-pin
checks and the live demo (sqlite I64 result round-trips through MkI64).

+++ examples/native/http_sqlite/native_shim.c
@@ -56,7 +56,7 @@ void demo_sqlite_close(void *db) { sqlite3_close((sqlite3 *)db); }
 
 /* ── POSIX TCP (CtHandle = fd cast to void*; CtInt64 port; CtString io) ─ */
 
-void *demo_tcp_listen(int64_t port) {
+void *demo_tcp_listen(int port) {
   int fd = socket(AF_INET, SOCK_STREAM, 0);
   if (fd < 0) shim_fail("socket() failed");
   int one = 1;
